Refactor TunicPortals into ERData and ERScripts
