+++ TunicRandomizer.csproj
@@ -216,7 +216,8 @@
     <Compile Include="src\Patches\SceneLoaderPatches.cs" />
     <Compile Include="src\Patches\SecretMayor.cs" />
     <Compile Include="src\Patches\TextBuilderPatches.cs" />
-    <Compile Include="src\Patches\TunicPortals.cs" />
+    <Compile Include="src\Patches\ERData.cs" />
+    <Compile Include="src\Patches\ERScripts.cs" />
     <Compile Include="src\Patches\SpeedrunFinishlineDisplayPatches.cs" />
     <Compile Include="src\Patches\SwordProgression.cs" />
     <Compile Include="src\Patches\TitleVersion.cs" />
