+++ src/Data/PortalCombo.cs
@@ -16,11 +16,11 @@ namespace TunicRandomizer {
         public Dictionary<string, int> AddComboRegion(Dictionary<string, int> inventory) {
             if (inventory.ContainsKey(Portal1.Region)) {
                 // for shop portals that are not in standalone
-                if (!TunicPortals.RegionDict.ContainsKey(Portal2.Region)) {
+                if (!ERData.RegionDict.ContainsKey(Portal2.Region)) {
                     inventory[Portal2.Region] = 1;
                     return inventory;
                 }
-                string outletRegion = TunicPortals.RegionDict[Portal2.Region].OutletRegion ?? Portal2.Region;
+                string outletRegion = ERData.RegionDict[Portal2.Region].OutletRegion ?? Portal2.Region;
                 inventory[outletRegion] = 1;
             }
             return inventory;
@@ -28,10 +28,10 @@ namespace TunicRandomizer {
 
         // check if this portal is oriented such that we should use the left-to-right shop look
         public bool FlippedShop() {
-            if (Portal1.Name.StartsWith("Shop Portal") && (Portal2.Direction == (int)TunicPortals.PDir.EAST || Portal2.Direction == (int)TunicPortals.PDir.SOUTH || Portal1.Direction == (int)TunicPortals.PDir.WEST)) {
+            if (Portal1.Name.StartsWith("Shop Portal") && (Portal2.Direction == (int)ERData.PDir.EAST || Portal2.Direction == (int)ERData.PDir.SOUTH || Portal1.Direction == (int)ERData.PDir.WEST)) {
                 return true;
             }
-            if (Portal2.Name.StartsWith("Shop Portal") && (Portal1.Direction == (int)TunicPortals.PDir.EAST || Portal1.Direction == (int)TunicPortals.PDir.SOUTH || Portal2.Direction == (int)TunicPortals.PDir.WEST)) {
+            if (Portal2.Name.StartsWith("Shop Portal") && (Portal1.Direction == (int)ERData.PDir.EAST || Portal1.Direction == (int)ERData.PDir.SOUTH || Portal2.Direction == (int)ERData.PDir.WEST)) {
                 return true;
             }
             return false;
